pyproject.toml
[project]
name = "clinic-ms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi-users[sqlalcemy]>=15.0.5",
    "fastapi[standard]>=0.136.0",
    "greenlet>=3.4.0",
    "pwdlib>=0.3.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlmodel>=0.0.38",
]
